requirements.txt
py==1.10.0
PyMySQL
allure_python_commons
pytest==5.4.0
requests_toolbelt==0.9.1
requests
pycryptodome==3.9.9
PyYAML
selenium==3.141.0
pytest-metadata
pytest-html
allure-pytest
jsonpath
redis
redis-py-cluster
binascii
hashlib
airtest
pocoui
